requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pycparser==2.21
PyNFe==0.4.4
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-simplexml==0.1.5
pytz==2024.1
requests==2.31.0
signxml==3.2.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
